+++ Changes.md
@@ -1,3 +1,8 @@
+* #715 - refactor database init in IntegrationTestUtils to factor out duplicate code
+* #740 - Validate rubric ranges when parsing a challenge description
+* #1324 - Add weighted rank to Championship category in finalist scheduling
+* #1315 - Fixed load error with finalist scheduling
+* #1323 - Better handling of display remote control errors 
 * #1303 - Track and display last page access for users
 * #1282 - Rework database schema to not be dependent on the challenge description
 * #1117 - Upgrade database on software startup. Most users won't see any effect from this.